using BigDecimal instead of Double
Former-commit-id: 901c4a95dcb150e5a1508721cacdf3150cdaad78

+++ api/src/main/scala/hr/com/blanka/apartments/query/price/QueryPriceRangeActor.scala
@@ -19,7 +19,7 @@ object QueryPriceRangeActor {
   def apply(dailyPriceActor: ActorRef) = Props(classOf[QueryPriceRangeActor], dailyPriceActor)
 }
 
-case class CalculationData(singleDayCalculations: Map[Long, Option[Double]])
+case class CalculationData(singleDayCalculations: Map[Long, Option[BigDecimal]])
 
 class QueryPriceRangeActor(dailyPriceActor: ActorRef) extends Actor {
